Add fields management and navigation
The form is now functional, but it's not very useful if we can't
navigate between fields.

Add some logic to move between fields when the user presses the `Enter`
key or clicks the `Continue` button. Also, add functionality to `Cancel`
button. If the user clicks the `Cancel` button, the form should reset
to the first field.
Finally, add some mock fields to test the navigation.