Add textarea
Currently, the field component only supports input elements. We need to add
support for textarea elements as well. The textarea element is very similar
to the input element. It has a label, a placeholder and a value, and it has
to animate in the same way as the input element, so we can reuse the same
`input-field` component.

Renames the `input-field` component to `field` and adds a `native-textarea`.
The `input` component now decides whether to render an input or a textarea.